Add exception handling to synchronize
If an exception is thrown in a synchronize() call, the future will now
still be set. synchronize() now returns a `bool` for void functors or
`std::optional<Re>` for functors specified to return `Re`.

This means that it's now possible to check the return value to see if
the synchronized call was made without throwing.

Signed-off-by: Ted Lyngmo <ted@lyncon.se>