Make delay functionality optional
Added a template parameter, `bool SetDelayEnabled = true`, that can be
set to `false` to disable the constructor taking `const duration& now_delay`
and the `set_delay_until(const time_point&)` member function.

If the delay functionality is not needed, this will make the queue
slightly more lightweight and the internal processing will be slightly
more effective.

Signed-off-by: Ted Lyngmo <ted@lyncon.se>